Split the parsers and worked on queries.
The parser module was split into three parts,

* model,
* query, and
* types.

Further work was done on the queries, which lead to alot of changes in
the modules which use the query parser and the query result.

The query will now be first parsed into an AST similar to the model and
later compiled to interpreter::Operations.